+++ client/src/types/myCollections/ICreateCollectionFormData.ts
@@ -1,0 +1,6 @@
+export interface ICreateCollectionFormData {
+  name: string
+  classification: string
+  language: string
+  defaultCollection: boolean
+}